% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasGnuMake.R
\name{hasGnuMake}
\alias{hasGnuMake}
\title{Check if GNU Make is available via the 'make' command}
\usage{
hasGnuMake()
}
\value{
Logical value indicating if GNU Make is available
}
\description{
Function checks if GNU Make is installed and available in the system PATH
via the 'make' command. It also verifies that the version of GNU Make is
at least the minimum required version needed by the package, which is
currently set to 3.82.
}
\examples{
if (hasGnuMake()) {
  message("GNU Make is available")
}
}
\author{
Michal Burda
}
