% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inShell.R
\name{inShell}
\alias{inShell}
\title{Convert R code to the character vector of shell commands evaluating the given R code.}
\usage{
inShell(...)
}
\arguments{
\item{...}{R commands to be converted}
}
\value{
A character vector of shell commands, which send the given R code by pipe to the R
interpreter
}
\description{
The function takes R commands, deparses them, substitutes existing variables, and converts
everything to character strings, from which a shell command is created that sends the given
R code to the R interpreter. Function is used internally to print the commands of R recipes
into `Makefile`.
}
\seealso{
[rRecipe()], [markdownRecipe()]
}
\author{
Michal Burda
}
