% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.recipe.R
\name{is.recipe}
\alias{is.recipe}
\title{Check if the argument is a valid recipe object.}
\usage{
is.recipe(x)
}
\arguments{
\item{x}{An argument to be tested}
}
\value{
`TRUE` if `x` is a valid recipe object and `FALSE` otherwise.
}
\description{
Function tests whether `x` is a valid recipe object, i.e., whether
it is list a list and inherits from the `recipe` S3 class. Instances
of `recipe` represent an atomic building unit, i.e. a command that
has to be executed, which optionally depends on some files or other
recipes -- see [recipe()] for more details.
}
\seealso{
[recipe()], [makefile()], [rRecipe()], [markdownRecipe()], [offlineRecipe()]
}
\author{
Michal Burda
}
