% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.rule.R
\name{is.rule}
\alias{is.rule}
\title{Check if the argument is a valid rule object.}
\usage{
is.rule(x)
}
\arguments{
\item{x}{An argument to be tested}
}
\value{
`TRUE` if `x` is a valid rule object and `FALSE` otherwise.
}
\description{
Function tests whether `x` is a valid rule object, i.e., whether
it is list a list and inherits from the `rmake.rule` S3 class. Instances
of `rule` represent an atomic building unit, i.e. a command that
has to be executed, which optionally depends on some files or other
rules -- see [rule()] for more details.
}
\seealso{
[rule()], [makefile()], [rRule()], [markdownRule()], [offlineRule()]
}
\author{
Michal Burda
}
