% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makerSkeleton.R
\name{makerSkeleton}
\alias{makerSkeleton}
\title{Prepare existing project for building with *maker*.}
\usage{
makerSkeleton()
}
\description{
This function creates a `Makefile.R` with an empty *maker* project
and generates a basic `Makefile` from it.
}
\seealso{
[makefile()], [recipe()]
}
\author{
Michal Burda
}
