% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offlineRecipe.R
\name{offlineRecipe}
\alias{offlineRecipe}
\title{Recipe for requesting manual user action}
\usage{
offlineRecipe(target, message, depends = NULL, task = "all")
}
\arguments{
\item{target}{A character vector of target file names of the manual (offline) build
command}

\item{message}{An error message to be issued if targets are older than dependencies
from `depends`}

\item{depends}{A character vector of file names the targets depend on}

\item{task}{A character vector of parent task names. The mechanism of tasks allows to
group recipes. Anything different from `'all'` will
cause creation of a new task depending on the given recipe. Executing `make taskname`
will then force building of this recipe.}
}
\value{
Instance of S3 class `recipe`
}
\description{
Instead of building the target, this recipe simply issues the given error message.
This recipe is useful for cases, where the target `target` depends on `depends`, but
has to be updated by some manual process. So if `target` is older than any of its
dependencies, `make` will throw an error until the user manualy updates the target.
}
\seealso{
[recipe()], [makefile()]
}
\author{
Michal Burda
}
