% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonRule.R
\name{pythonRule}
\alias{pythonRule}
\title{Rule for running Python scripts}
\usage{
pythonRule(
  target,
  script,
  depends = NULL,
  args = NULL,
  python = "python",
  task = "all"
)
}
\arguments{
\item{target}{Name of output files to be created}

\item{script}{Name of the Python script to be executed}

\item{depends}{A vector of file names that the Python script depends on, or
\code{NULL}.}

\item{args}{A string with command line arguments for the Python script.
This string may contain variables \verb{$[DEP1]}, \verb{$[DEP2]}, ..., \verb{$[TARGET1]},
\verb{$[TARGET2]}, ... that are replaced with the corresponding file names from
\code{depends} and \code{target}. If \code{args} is \code{NULL}, the arguments are created
from the \code{depends} and \code{target} parameters, in that order.}

\item{python}{The command to execute Python. By default, it is \code{python}, but
it can be set to a specific path or to \code{python3} if needed.}

\item{task}{A character vector of parent task names. The mechanism of tasks allows
grouping rules. Anything different from \code{'all'} will cause the creation of a
new task depending on the given rule. Executing \verb{make taskname} will then
force building this rule.}
}
\value{
Instance of S3 class \code{rmake.rule}
}
\description{
This rule executes Python scripts to create various file outputs.
}
\details{
In detail, this rule executes the following command in a shell:
\verb{python script arguments}
where \code{arguments} are script arguments created from the \code{args} parameter. If
\code{args} is \code{NULL}, the arguments are created from the \code{depends} and \code{target}
parameters, in that order. If \code{args} is not \code{NULL}, it must be a string
with command line arguments that may contain variables \verb{$[DEP1]}, \verb{$[DEP2]},
..., \verb{$[TARGET1]}, \verb{$[TARGET2]}, ... that are replaced with the corresponding
file names from \code{depends} and \code{target}.

Issuing \verb{make clean} from the shell causes removal of all files specified in
the \code{target} parameter.
}
\seealso{
\code{\link[=rule]{rule()}}, \code{\link[=makefile]{makefile()}}
}
\author{
Michal Burda
}
