% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{recipe}
\alias{recipe}
\title{General creator of an instance of the S3 `recipe` class}
\usage{
recipe(target, depends = NULL, build = NULL, clean = NULL, task = "all")
}
\arguments{
\item{target}{A character vector of target file names that are created by the given build command}

\item{depends}{A character vector of file names the build command depends on}

\item{build}{A shell command that runs the build of the given target}

\item{clean}{A shell command that erases all files produced by the build command}

\item{task}{A character vector of parent task names. The mechanism of tasks allows to
group recipes. Anything different from `'all'` will
cause creation of a new task depending on the given recipe. Executing `make taskname`
will then force building of this recipe.}
}
\value{
Instance of S3 class `recipe`
}
\description{
Recipe is an atomic element of the build process. It defines a set of `target` file names,
which are to be built with a given `build` command from a given set `depends` of files
that targets depend on, and which can be removed by a given `clean` command.
}
\details{
If there is a need to group some recipes together, one can assign them the same task identifier in
the `task` argument. Each recipe may get assigned one or more tasks. Tasks may be then built
by executing `make task_name` on the command line, which forces to rebuild all recipes assigned to the
task `'task_name'`. By default, all recipes are assigned to task `all`,
which causes `make all` command to build everything.
}
\seealso{
[makefile()]
}
\author{
Michal Burda
}
