% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmake.R
\name{rmake_path_dependencies}
\alias{rmake_path_dependencies}
\title{List the path depedencies found by rmake}
\usage{
rmake_path_dependencies(project = NULL)
}
\arguments{
\item{project}{A path to a folder, default is NULL. If provided, \code{rmake} will consider
the path given by \code{project} as the active project. It will then apply to all R files from
this project (not to the current one).}
}
\value{
It returns a list. The names of the list are the names of the code chunks in the
following format: \code{filename@chunkname}. Each element of the list contains
two elements: \code{input} and \code{output}, which are both character vectors of paths.
}
\description{
Use this function to display the path dependencies found by rmake across
all the chunks from the project.
}
