% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmakeSkeleton.R
\name{rmakeSkeleton}
\alias{rmakeSkeleton}
\title{Prepare existing project for building with *rmake*.}
\usage{
rmakeSkeleton()
}
\description{
This function creates a `Makefile.R` with an empty *rmake* project
and generates a basic `Makefile` from it.
}
\seealso{
[makefile()], [rule()]
}
\author{
Michal Burda
}
