% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRecipe.R
\name{rRecipe}
\alias{rRecipe}
\title{Recipe for running R scripts}
\usage{
rRecipe(target, script, depends = NULL, params = NULL, task = "all")
}
\arguments{
\item{target}{Name of output files to be created}

\item{script}{Name of the R script to be executed}

\item{depends}{A vector of file names that the R script depends on, or `NULL`.}

\item{params}{`NULL` or a list of R values that become available within the `script` in
a `params` variable.}

\item{task}{A character vector of parent task names. The mechanism of tasks allows to
group recipes. Anything different from `'all'` will
cause creation of a new task depending on the given recipe. Executing `make taskname`
will then force building of this recipe.}
}
\value{
Instance of S3 class `recipe`
}
\description{
This recipe is for execution of R scripts in order to create various file outputs.
}
\details{
In detail, this recipe executes the following command in a separate R process:
`params <- params; source(script)`

Issuing `make clean` from the shell causes removal of all files specified in `target` parameter.
}
\seealso{
[recipe()], [makefile()], [markdownRecipe()]
}
\author{
Michal Burda
}
