% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizePath.R
\name{sanitizePath}
\alias{sanitizePath}
\title{Sanitize a file path for the current operating system}
\usage{
sanitizePath(path)
}
\arguments{
\item{path}{A character string representing the file path to be sanitized.}
}
\value{
A sanitized file path suitable for the current operating system.
}
\description{
This function replaces forward slashes with backslashes on Windows systems,
and leaves the path unchanged on Unix-like systems.
}
\author{
Michal Burda
}
