% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeSpaces.R
\name{sanitizeSpaces}
\alias{sanitizeSpaces}
\title{Escape spaces in a string as needed in file names used in Makefile files}
\usage{
sanitizeSpaces(x)
}
\arguments{
\item{x}{A character vector to be sanitized}
}
\value{
A character vector with spaces replaced by \verb{\\ }
}
\description{
Escape spaces in a string as needed in file names used in Makefile files
}
\author{
Michal Burda
}
