% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile.R
\docType{data}
\name{defaultVars}
\alias{defaultVars}
\title{Variables used within Makefile generating process}
\format{
An object of class \code{character} of length 4.
}
\usage{
defaultVars
}
\description{
\code{defaultVars} is a reserved variable, a named vector that defines
Makefile variables, i.e. shell variables that will exist during
the execution of Makefile rules. The content of this variable
is written into the resulting Makefile within the execution of
the \code{\link[=makefile]{makefile()}} function.
}
\seealso{
\code{\link[=makefile]{makefile()}}
}
\author{
Michal Burda
}
\keyword{datasets}
