% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasGnuMake.R
\name{hasGnuMake}
\alias{hasGnuMake}
\title{Check if GNU Make is available via the 'make' command}
\usage{
hasGnuMake()
}
\value{
Logical value indicating if GNU Make is available
}
\description{
Check if GNU Make is available via the 'make' command
}
\examples{
if (hasGnuMake()) {
  message("GNU Make is available")
}
}
\author{
Michal Burda
}
