% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasGnuMake.R
\name{hasGnuMake}
\alias{hasGnuMake}
\title{Check if GNU Make is available via the 'make' command}
\usage{
hasGnuMake(version = "3.82")
}
\arguments{
\item{version}{Optional minimum version string to check (the default value,
i.e., "3.82" is the minimum required version for rmake package functionality)}
}
\value{
Logical value indicating if GNU Make is available
}
\description{
Check if GNU Make is available via the 'make' command
}
\examples{
if (hasGnuMake()) {
  message("GNU Make is available")
}
}
\author{
Michal Burda
}
