% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.rule.R
\name{is.rule}
\alias{is.rule}
\title{Check if the argument is a valid rule object.}
\usage{
is.rule(x)
}
\arguments{
\item{x}{An argument to be tested}
}
\value{
\code{TRUE} if \code{x} is a valid rule object and \code{FALSE} otherwise.
}
\description{
Function tests whether \code{x} is a valid rule object, i.e., whether
it is list a list and inherits from the \code{rmake.rule} S3 class. Instances
of \code{rule} represent an atomic building unit, i.e. a command that
has to be executed, which optionally depends on some files or other
rules -- see \code{\link[=rule]{rule()}} for more details.
}
\seealso{
\code{\link[=rule]{rule()}}, \code{\link[=makefile]{makefile()}}, \code{\link[=rRule]{rRule()}}, \code{\link[=markdownRule]{markdownRule()}}, \code{\link[=offlineRule]{offlineRule()}}
}
\author{
Michal Burda
}
