% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\title{Run `make`` in the system}
\usage{
make(...)
}
\arguments{
\item{...}{Command-line arguments passed to the \code{make} command}
}
\value{
Exit status of the command, see \code{\link[base:system2]{base::system2()}} for details.
}
\description{
This function executes the \code{make} command in order to re-build all dependencies, accordingly to \code{Makefile}
generated by \code{\link[=makefile]{makefile()}}.
}
\examples{
\dontrun{
  make()        # make all
  make('clean') # make the 'clean' task
  make('-j', 4) # make with 4 processes in parallell
}
}
\seealso{
\code{\link[=makefile]{makefile()}}, \code{\link[=rmakeSkeleton]{rmakeSkeleton()}}
}
\author{
Michal Burda
}
