% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownRecipe.R
\name{markdownRecipe}
\alias{markdownRecipe}
\title{Recipe for building text documents from Markdown files}
\usage{
markdownRecipe(target, script, depends = NULL, format = "all",
  params = NULL, task = "all")
}
\arguments{
\item{target}{Name of the output file to be created}

\item{script}{Name of the markdown file to be rendered}

\item{depends}{A vector of file names that the markdown script depends on.}

\item{format}{Requested format of the result. Parameter is passed as `format` argument
to [markdown::render()]. Allowed values are: 'all', 'html_document', 'pdf_document',
'word_document', 'odt_document', 'rtf_document', or 'md_document'.}

\item{params}{A list of R values that become available within the `script` in variables of names
corresponding to names of elements in the list.}
}
\value{
Instance of S3 class `recipe`
}
\description{
This recipe is for execution of Markdown rendering in order to create text file
of various supported formats such as (PDF, DOCX, etc.).
}
\seealso{
[recipe()], [makefile()], [rRecipe()]
}
\author{
Michal Burda
}
