% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownRule.R
\name{markdownRule}
\alias{markdownRule}
\title{Rule for building text documents from Markdown files}
\usage{
markdownRule(target, script, depends = NULL, params = list(), task = "all")
}
\arguments{
\item{target}{Name of the output file to be created}

\item{script}{Name of the markdown file to be rendered}

\item{depends}{A vector of file names that the markdown script depends on, or \code{NULL}.}

\item{params}{A list of R values that become available within the \code{script} in
a \code{params} variable.}

\item{task}{A character vector of parent task names. The mechanism of tasks allows
grouping rules. Anything different from \code{'all'} will
cause the creation of a new task depending on the given rule. Executing \verb{make taskname}
will then force building this rule.}
}
\value{
Instance of S3 class \code{rmake.rule}
}
\description{
This rule executes Markdown rendering to create text files
in various supported formats such as PDF, DOCX, etc.
}
\details{
This rule executes the following command in a separate R process:
\verb{params <- params; rmarkdown::render(script, output_format=format, output_file=target)}

That is, the parameters given in the \code{params} argument are stored in the global variable
and then the \code{script} is rendered with rmarkdown. Note that the re-generation of the
\code{Makefile} with any change to \code{params} will not cause the re-execution of the recipe unless
other script dependencies change.

Issuing \verb{make clean} from the shell causes removal of all files specified in the \code{target} parameter.
}
\examples{
r <- markdownRule(target='report.pdf',
                  script='report.Rmd',
                  depends=c('data1.csv', 'data2.csv'))

# generate the content of a makefile (as character vector)
makefile(list(r))

# generate to file
tmp <- tempdir()
makefile(list(r), file.path(tmp, "Makefile"))
}
\seealso{
\code{\link[=rule]{rule()}}, \code{\link[=makefile]{makefile()}}, \code{\link[=rRule]{rRule()}}
}
\author{
Michal Burda
}
